Wrap charters and README to 80 columns
There are, I would argue, two reasonable choices for wrapping text
files.  One is to not hard wrap lines at all.  This has some
advantages and is the current Reference policy.  The other is to hard
wrap lines consistently at some column.  This is what the FLS does, at
80 columns.  This too has some advantages.

For the moment, let's be consistent with the FLS itself and hard wrap
our own documents at 80 columns consistently.