Revise README
Verbiage that talks about it being time to create a team is going to
age quickly.  Let's just talk about the purpose of the team and link
to the charter.

Though the Project is inconsistent about this, let's err on the formal
side (given our mission) and consistently capitalize "team" in "the
FLS Team".