Use `--` as separator in headers
What we really want as a separator for the headers is an em dash.  We
could use the Unicode character, but that tends to not look very good
when rendered in fixed width fonts, and much of the time the file will
be viewed this way.  We could use `&mdash;`, but that also presents
poorly when not rendered.  What might be best would be if GitHub
rendered `---` as an em dash, as some Markdown renderers do, but alas,
they don't do this.

Let's use `--` as the header separator.  GH will just render this as
two dashes, but that still looks OK, and it looks OK when not rendered
as well.